space rock hath fallen
here I was siting and pondering the meaning of the universe when before my eyes a flash of light in the sky lit up thine world. a blast so great night was cast away in favor of day. a true sight to behold in deed. but that taws not they end for a bang was heard in the near by forest. a flowing rock shining red in the dirt. I had to touché it. I had to. but that was how thine lost his hand. a sad and painful loss.